requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
idna==3.11
keyboard==0.13.5
motor==3.6.0
mss==10.1.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
